Move `mock/values.ts` to `mock-values.ts`
