+++ packages/lib/src/index.ts
@@ -198,7 +198,7 @@ export type {
 } from './interactions/models';
 
 // Mock data and utilities
-export { mockValues } from '@h5web/shared/mock/values';
+export { mockValues } from '@h5web/shared/mock-values';
 export { makeMockFile } from '@h5web/shared/mock/metadata';
 export { findMockEntity } from '@h5web/shared/mock/utils';
