+++ packages/shared/package.json
@@ -18,7 +18,7 @@
     "./mock/metadata": "./src/mock/metadata.ts",
     "./mock/models": "./src/mock/models.ts",
     "./mock/utils": "./src/mock/utils.ts",
-    "./mock/values": "./src/mock/values.ts",
+    "./mock-values": "./src/mock-values.ts",
     "./styles.css": "./src/styles.css",
     ".": "./src/index.ts"
   },
